+++ implementacao/server/migrations/database_update4.py
@@ -1,10 +1,21 @@
 import sqlite3
+import sys
 
-con = sqlite3.connect('war.db')
+wardb = 'war.db'
+if len(sys.argv) > 1:
+    wardb = sys.argv[1]
 
+con = sqlite3.connect(wardb)
 with con:
     cur = con.cursor()
-    cur.execute("DROP TABLE IF EXISTS GrupoUsuarios")
-    cur.execute("CREATE TABLE GrupoUsuarios(id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, idUsuario INTEGER, nome TEXT)")
-    cur.execute("INSERT INTO Versao(sistema, versao) VALUES ('db', '4')")
-    con.commit()
+    row = cur.execute("SELECT versao FROM Versao WHERE sistema = 'db' ORDER BY versao DESC LIMIT 1").fetchone()
+    if row and row[0] == '3':
+        print('Aplicando Patch database_update4...')
+        cur.execute("DROP TABLE IF EXISTS GrupoUsuarios")
+        cur.execute(
+            "CREATE TABLE GrupoUsuarios(id INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL, idUsuario INTEGER, nome TEXT)")
+        cur.execute("INSERT INTO Versao(sistema, versao) VALUES ('db', '4')")
+        con.commit()
+        print('Patch database_update4 aplicado com sucesso.')
+    else:
+        print('Patch database_update4 já foi aplicado anteriormente.')
